Responsive design fixed
Fixed the issues with responsive design. Fixed other format issues as well (just moving items and controlling spaces, nothing that big).

Changed my Hobby page with a new design. Feel free to use this design for your hobby page if you want.

Any other issues are fixed as well. If something appears, let us know in the team chat!

I was thinking on adding an Admin version to the site but if time lets me do that, I'll add it soon.